물웅덩이: 조건 및 예제 수정

+++ mo/kmo-2021-high-euler-1st-19/solution-easy.cpp
@@ -49,7 +49,7 @@ int main()
     t_start = system_clock::now();
 
     stack.push(make_tuple(start.first, start.second, 0));
-    int ans = 0;
+    long long ans = 0;
 
     while(!stack.empty()) {
         int i, j, cnt;
